gen_src/requirements.txt
fastapi~=0.111.0
uvicorn[standard]~=0.29.0
sqlalchemy~=2.0.0
pydantic~=2.7.0
pydantic-settings~=2.2.0
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.0
aiohttp~=3.9.0
pandas~=2.2.0
TA-Lib~=0.4.0
google-cloud-aiplatform~=1.49.0
APScheduler~=3.10.0
python-dotenv~=1.0.0
psycopg2-binary~=2.9.0